addressed review comments - deleted a new Router Record Strategy
