+++ data-prepper-core/src/main/java/org/opensearch/dataprepper/pipeline/Pipeline.java
@@ -18,7 +18,6 @@ import org.opensearch.dataprepper.pipeline.common.PipelineThreadFactory;
 import org.opensearch.dataprepper.pipeline.common.PipelineThreadPoolExecutor;
 import org.opensearch.dataprepper.pipeline.router.Router;
 import org.opensearch.dataprepper.pipeline.router.RouterCopyRecordStrategy;
-import org.opensearch.dataprepper.pipeline.router.RouterAcknowledgementsRecordStrategy;
 import org.opensearch.dataprepper.pipeline.router.RouterGetRecordStrategy;
 import org.opensearch.dataprepper.model.event.EventFactory;
 import org.opensearch.dataprepper.model.acknowledgements.AcknowledgementSetManager;
@@ -286,9 +285,9 @@ public class Pipeline {
     List<Future<Void>> publishToSinks(final Collection<Record> records) {
         final int sinksSize = sinks.size();
         final List<Future<Void>> sinkFutures = new ArrayList<>(sinksSize);
-        final RouterGetRecordStrategy getRecordStrategy = (source.areAcknowledgementsEnabled()) ? 
-                    new RouterAcknowledgementsRecordStrategy(eventFactory, acknowledgementSetManager, sinks) :
-                    new RouterCopyRecordStrategy(sinks);
+
+        final RouterGetRecordStrategy getRecordStrategy =
+                    new RouterCopyRecordStrategy(eventFactory, (source.areAcknowledgementsEnabled()) ? acknowledgementSetManager : null, sinks);
         router.route(records, sinks, getRecordStrategy, (sink, events) ->
                 sinkFutures.add(sinkExecutorService.submit(() -> sink.output(events), null))
         );
