+++ data-prepper-core/src/main/java/org/opensearch/dataprepper/pipeline/router/RouterCopyRecordStrategy.java
@@ -14,6 +14,12 @@ import org.opensearch.dataprepper.model.trace.JacksonSpan;
 import org.opensearch.dataprepper.model.record.Record;
 import org.opensearch.dataprepper.model.event.Event;
 import org.opensearch.dataprepper.model.event.JacksonEvent;
+import org.opensearch.dataprepper.model.event.EventFactory;
+import org.opensearch.dataprepper.model.event.EventHandle;
+import org.opensearch.dataprepper.model.event.EventBuilder;
+import org.opensearch.dataprepper.model.event.EventMetadata;
+import org.opensearch.dataprepper.event.DefaultEventHandle;
+import org.opensearch.dataprepper.model.acknowledgements.AcknowledgementSetManager;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -23,9 +29,15 @@ import java.util.HashSet;
 
 public class RouterCopyRecordStrategy implements RouterGetRecordStrategy {
     private Set<Record> routedRecords;
+    private Set<Record> referencedRecords;
+    private AcknowledgementSetManager acknowledgementSetManager;
+    private EventFactory eventFactory;
 
-    public <C> RouterCopyRecordStrategy(final Collection<DataFlowComponent<C>> dataFlowComponents) {
+    public <C> RouterCopyRecordStrategy(final EventFactory eventFactory, final AcknowledgementSetManager acknowledgementSetManager, final Collection<DataFlowComponent<C>> dataFlowComponents) {
+        this.acknowledgementSetManager = acknowledgementSetManager;
+        this.eventFactory = eventFactory;
         routedRecords = null;
+        referencedRecords = new HashSet<Record>();
         /*
          * If there are more than one sink and one of the sinks is
          * pipeline connector, then we should make a copy of every
@@ -42,12 +54,31 @@ public class RouterCopyRecordStrategy implements RouterGetRecordStrategy {
         }
     }
 
+    Set<Record> getReferencedRecords() {
+        return referencedRecords;
+    }
+
+    private void acquireEventReference(Record record) {
+        if (acknowledgementSetManager == null || record.getData() == null) {
+            return;
+        }
+        if (referencedRecords.contains(record) || ((routedRecords != null) && routedRecords.contains(record))) {
+            EventHandle eventHandle = ((JacksonEvent)record.getData()).getEventHandle();
+            if (eventHandle != null) {
+                acknowledgementSetManager.acquireEventReference(eventHandle);
+            }
+        } else if (!referencedRecords.contains(record)) {
+            referencedRecords.add(record);
+        }
+    }
     @Override
     public Record getRecord(Record record) {
         if (routedRecords == null) {
+            acquireEventReference(record);
             return record;
         }
         if (!routedRecords.contains(record)) {
+            acquireEventReference(record);
             routedRecords.add(record);
             return record;
         }
@@ -64,8 +95,21 @@ public class RouterCopyRecordStrategy implements RouterGetRecordStrategy {
                 final Event recordEvent = (Event) record.getData();
                 JacksonEvent newRecordEvent;
                 Record newRecord;
-                newRecordEvent = JacksonEvent.fromEvent(recordEvent);
-                newRecord = new Record<>(newRecordEvent);
+                DefaultEventHandle eventHandle = (DefaultEventHandle)recordEvent.getEventHandle();
+                if (eventHandle != null) {
+                    final EventMetadata eventMetadata = recordEvent.getMetadata();
+                    final EventBuilder eventBuilder = (EventBuilder) eventFactory.eventBuilder(EventBuilder.class).withEventMetadata(eventMetadata).withData(recordEvent.toMap());
+                    newRecordEvent = (JacksonEvent) eventBuilder.build();
+
+                    eventHandle.getAcknowledgementSet().add(newRecordEvent);
+                    newRecord = new Record<>(newRecordEvent);
+                    acquireEventReference(newRecord);
+                } else {
+                    // TODO we should have a way to create from factory
+                    // even when acknowledgements are not used
+                    newRecordEvent = JacksonEvent.fromEvent(recordEvent);
+                    newRecord = new Record<>(newRecordEvent);
+                }
                 return newRecord;
             } catch (Exception ex) {
             }
@@ -76,12 +120,14 @@ public class RouterCopyRecordStrategy implements RouterGetRecordStrategy {
     @Override
     public Collection<Record> getAllRecords(final Collection<Record> allRecords) {
         if (routedRecords == null) {
+            allRecords.stream().forEach((record) -> acquireEventReference(record));
             return allRecords;
         }
         if (routedRecords.isEmpty()) {
+            allRecords.stream().forEach((record) -> acquireEventReference(record));
             routedRecords.addAll(allRecords);
             return allRecords;
-        } 
+        }
         List<Record> newRecords = new ArrayList<Record>();
         for (Record record : allRecords) {
             newRecords.add(getRecord(record));
