+++ data-prepper-core/src/test/java/org/opensearch/dataprepper/pipeline/router/RouterCopyRecordStrategyTests.java
@@ -18,7 +18,10 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Map;
+import java.util.Iterator;
 import java.util.HashMap;
+import java.util.List;
+import java.util.ArrayList;
 import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -26,14 +29,25 @@ import java.util.stream.IntStream;
 
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.lenient;
+import static org.mockito.Mockito.doAnswer;
+import static org.mockito.ArgumentMatchers.any;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.CoreMatchers.sameInstance;
 import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.Assert.assertFalse;
 import org.opensearch.dataprepper.model.event.Event;
 import org.opensearch.dataprepper.model.event.JacksonEvent;
+import org.opensearch.dataprepper.model.event.EventFactory;
+import org.opensearch.dataprepper.model.event.EventHandle;
+import org.opensearch.dataprepper.model.event.EventBuilder;
+import org.opensearch.dataprepper.model.event.EventMetadata;
+import org.opensearch.dataprepper.event.DefaultEventHandle;
+import org.opensearch.dataprepper.model.acknowledgements.AcknowledgementSetManager;
+import org.opensearch.dataprepper.model.acknowledgements.AcknowledgementSet;
 
 @ExtendWith(MockitoExtension.class)
 public class RouterCopyRecordStrategyTests {
@@ -41,19 +55,62 @@ public class RouterCopyRecordStrategyTests {
     @Mock
     private DataFlowComponent<PipelineConnector> pipelineDataFlowComponent;
     private Collection<Record> recordsIn;
+    private Collection<Record> mockRecordsIn;
+    @Mock
+    private AcknowledgementSetManager acknowledgementSetManager;
+    @Mock
+    private AcknowledgementSet acknowledgementSet1;
+    @Mock
+    private EventFactory eventFactory;
+    @Mock
+    private EventBuilder eventBuilder;
+
+    private JacksonEvent event;
+
+    private Map<EventHandle, Integer> handleRefCount;
 
     private static class TestComponent {
     }
 
     @BeforeEach
     void setUp() {
-        recordsIn = IntStream.range(0, 10)
+        handleRefCount = new HashMap<>();
+        eventFactory = mock(EventFactory.class);
+        acknowledgementSetManager = mock(AcknowledgementSetManager.class);
+        acknowledgementSet1 = mock(AcknowledgementSet.class);
+        try {
+            lenient().doAnswer((i) -> {
+                EventHandle handle = (EventHandle) i.getArgument(0);
+                int v = handleRefCount.getOrDefault(handle, 0);
+                handleRefCount.put(handle, v+1);
+                return null;
+            }).when(acknowledgementSetManager).acquireEventReference(any(EventHandle.class));
+        } catch (Exception e){}
+        mockRecordsIn = IntStream.range(0, 10)
                 .mapToObj(i -> mock(Record.class))
+                .collect(Collectors.toList());
+        final Map<String, Object> eventMap = new HashMap<>();
+        eventMap.put(UUID.randomUUID().toString(), UUID.randomUUID().toString());
+        event = JacksonEvent.builder().withData(eventMap)
+                                      .withEventType("event")
+                                      .build();
+        recordsIn = IntStream.range(0, 10)
+                .mapToObj(i -> new Record<Event>(JacksonEvent.fromEvent(event)))
                 .collect(Collectors.toList());
     }
 
-    private <C> RouterGetRecordStrategy createObjectUnderTest(Collection<DataFlowComponent<C>> dataFlowComponents) {
-        return new RouterCopyRecordStrategy(dataFlowComponents);
+    private void attachEventHandlesToRecordsIn(List<EventHandle> eventHandles) {
+        Iterator iter = recordsIn.iterator();
+        while (iter.hasNext()) {
+            Record r = (Record) iter.next();
+            DefaultEventHandle handle = new DefaultEventHandle(acknowledgementSet1);
+            ((JacksonEvent)r.getData()).setEventHandle(handle);
+            eventHandles.add(handle);
+        }
+    }
+
+    private <C> RouterCopyRecordStrategy createObjectUnderTest(Collection<DataFlowComponent<C>> dataFlowComponents) {
+        return new RouterCopyRecordStrategy(eventFactory, acknowledgementSetManager, dataFlowComponents);
     }
 
     @Test
@@ -66,37 +123,37 @@ public class RouterCopyRecordStrategyTests {
         DataFlowComponent<TestComponent> dataFlowComponent = mock(DataFlowComponent.class);
         Collection<DataFlowComponent<TestComponent>> dataFlowComponents
             = Collections.singletonList(dataFlowComponent);
-        
-        final RouterGetRecordStrategy getRecordStrategy = createObjectUnderTest(dataFlowComponents);
-        Record firstRecord = recordsIn.iterator().next();
+
+        final RouterCopyRecordStrategy getRecordStrategy = createObjectUnderTest(dataFlowComponents);
+        Record firstRecord = mockRecordsIn.iterator().next();
         Record recordOut = getRecordStrategy.getRecord(firstRecord);
         assertThat(recordOut, sameInstance(firstRecord));
         recordOut = getRecordStrategy.getRecord(firstRecord);
         assertThat(recordOut, sameInstance(firstRecord));
-        Collection<Record> recordsOut = getRecordStrategy.getAllRecords(recordsIn);
-        assertThat(recordsOut.size(), equalTo(recordsIn.size()));
-        assertThat(recordsIn.stream().collect(Collectors.toSet()), equalTo(recordsOut.stream().collect(Collectors.toSet())));
+        Collection<Record> recordsOut = getRecordStrategy.getAllRecords(mockRecordsIn);
+        assertThat(recordsOut.size(), equalTo(mockRecordsIn.size()));
+        assertThat(mockRecordsIn.stream().collect(Collectors.toSet()), equalTo(recordsOut.stream().collect(Collectors.toSet())));
     }
 
     @Test
     void test_with_one_data_flow_pipeline_connector() {
         DataFlowComponent<PipelineConnector> dataFlowComponent = mock(DataFlowComponent.class);
-        
+
         Collection<DataFlowComponent<PipelineConnector>> dataFlowComponents
             = Collections.singletonList(dataFlowComponent);
-        final RouterGetRecordStrategy getRecordStrategy = createObjectUnderTest(dataFlowComponents);
-        Record firstRecord = recordsIn.iterator().next();
+        final RouterCopyRecordStrategy getRecordStrategy = createObjectUnderTest(dataFlowComponents);
+        Record firstRecord = mockRecordsIn.iterator().next();
         Record recordOut = getRecordStrategy.getRecord(firstRecord);
         assertThat(recordOut, sameInstance(firstRecord));
         recordOut = getRecordStrategy.getRecord(firstRecord);
         assertThat(recordOut, sameInstance(firstRecord));
-        Collection<Record> recordsOut = getRecordStrategy.getAllRecords(recordsIn);
-        assertThat(recordsOut.size(), equalTo(recordsIn.size()));
-        assertThat(recordsIn.stream().collect(Collectors.toSet()), equalTo(recordsOut.stream().collect(Collectors.toSet())));
+        Collection<Record> recordsOut = getRecordStrategy.getAllRecords(mockRecordsIn);
+        assertThat(recordsOut.size(), equalTo(mockRecordsIn.size()));
+        assertThat(mockRecordsIn.stream().collect(Collectors.toSet()), equalTo(recordsOut.stream().collect(Collectors.toSet())));
         /* Repeat the test to see if the conditions holds true */
-        recordsOut = getRecordStrategy.getAllRecords(recordsIn);
-        assertThat(recordsOut.size(), equalTo(recordsIn.size()));
-        assertThat(recordsIn.stream().collect(Collectors.toSet()), equalTo(recordsOut.stream().collect(Collectors.toSet())));
+        recordsOut = getRecordStrategy.getAllRecords(mockRecordsIn);
+        assertThat(recordsOut.size(), equalTo(mockRecordsIn.size()));
+        assertThat(mockRecordsIn.stream().collect(Collectors.toSet()), equalTo(recordsOut.stream().collect(Collectors.toSet())));
     }
 
     @Test
@@ -107,16 +164,7 @@ public class RouterCopyRecordStrategyTests {
             dataFlowComponents.add(pipelineDataFlowComponent);
         }
 
-        final Map<String, Object> eventMap = new HashMap<>();
-        eventMap.put(UUID.randomUUID().toString(), UUID.randomUUID().toString());
-        final JacksonEvent event = JacksonEvent.builder()
-                                       .withData(eventMap)
-                                       .withEventType("event")
-                                       .build();
-        recordsIn = IntStream.range(0, 10)
-                .mapToObj(i -> new Record<Event>(event))
-                .collect(Collectors.toList());
-        final RouterGetRecordStrategy getRecordStrategy = createObjectUnderTest(dataFlowComponents);
+        final RouterCopyRecordStrategy getRecordStrategy = createObjectUnderTest(dataFlowComponents);
         Record firstRecord = recordsIn.iterator().next();
         Record recordOut = getRecordStrategy.getRecord(firstRecord);
         assertThat(recordOut, sameInstance(firstRecord));
@@ -127,9 +175,140 @@ public class RouterCopyRecordStrategyTests {
         /* Repeat the test to make sure the records are copied second time */
         recordsOut = getRecordStrategy.getAllRecords(recordsIn);
         assertThat(recordsOut.size(), equalTo(recordsIn.size()));
-        
+
         Set<Record> recordsOutSet =  recordsOut.stream().collect(Collectors.toSet());
         recordsIn.forEach(recordIn -> assertFalse(recordsOutSet.contains(recordIn)));
     }
 
+    @Test
+    void test_one_record_with_acknowledgements() {
+        DataFlowComponent<TestComponent> dataFlowComponent = mock(DataFlowComponent.class);
+        Collection<DataFlowComponent<TestComponent>> dataFlowComponents
+            = Collections.singletonList(dataFlowComponent);
+
+        final RouterCopyRecordStrategy getRecordStrategy = createObjectUnderTest(dataFlowComponents);
+        List<EventHandle> eventHandles = new ArrayList<>();
+        attachEventHandlesToRecordsIn(eventHandles);
+        Record firstRecord = recordsIn.iterator().next();
+        EventHandle firstHandle = ((Event)firstRecord.getData()).getEventHandle();
+        Record recordOut = getRecordStrategy.getRecord(firstRecord);
+        assertThat(recordOut, sameInstance(firstRecord));
+        assertTrue(getRecordStrategy.getReferencedRecords().contains(firstRecord));
+        recordOut = getRecordStrategy.getRecord(firstRecord);
+        assertThat(recordOut, sameInstance(firstRecord));
+        assertThat(handleRefCount.get(firstHandle), equalTo(1));
+        recordOut = getRecordStrategy.getRecord(firstRecord);
+        assertThat(recordOut, sameInstance(firstRecord));
+        assertThat(handleRefCount.get(firstHandle), equalTo(2));
+    }
+
+    @Test
+    void test_multiple_records_with_acknowledgements() {
+        DataFlowComponent<TestComponent> dataFlowComponent = mock(DataFlowComponent.class);
+        Collection<DataFlowComponent<TestComponent>> dataFlowComponents
+            = Collections.singletonList(dataFlowComponent);
+
+        final RouterCopyRecordStrategy getRecordStrategy = createObjectUnderTest(dataFlowComponents);
+        List<EventHandle> eventHandles = new ArrayList<>();
+        attachEventHandlesToRecordsIn(eventHandles);
+        Collection<Record> recordsOut = getRecordStrategy.getAllRecords(recordsIn);
+        assertThat(recordsOut.size(), equalTo(recordsIn.size()));
+        assertThat(recordsIn.stream().collect(Collectors.toSet()), equalTo(recordsOut.stream().collect(Collectors.toSet())));
+        recordsIn.stream().forEach((record) -> {
+            assertTrue(getRecordStrategy.getReferencedRecords().contains(record));
+        });
+        recordsOut = getRecordStrategy.getAllRecords(recordsIn);
+        assertThat(recordsOut.size(), equalTo(recordsIn.size()));
+        assertThat(recordsIn.stream().collect(Collectors.toSet()), equalTo(recordsOut.stream().collect(Collectors.toSet())));
+        eventHandles.stream().forEach((handle) -> {
+            assertThat(handleRefCount.get(handle), equalTo(1));
+        });
+        recordsOut = getRecordStrategy.getAllRecords(recordsIn);
+        eventHandles.stream().forEach((handle) -> {
+            assertThat(handleRefCount.get(handle), equalTo(2));
+        });
+    }
+
+    @Test
+    void test_one_record_with_acknowledgements_and_multi_components() {
+        Collection<DataFlowComponent<PipelineConnector>> dataFlowComponents = new ArrayList<>();
+        when(pipelineDataFlowComponent.getComponent()).thenReturn(new PipelineConnector());
+        for (int i = 0; i < 3; i++) {
+            dataFlowComponents.add(pipelineDataFlowComponent);
+        }
+
+        final RouterCopyRecordStrategy getRecordStrategy = createObjectUnderTest(dataFlowComponents);
+        List<EventHandle> eventHandles = new ArrayList<>();
+        attachEventHandlesToRecordsIn(eventHandles);
+        try {
+            doAnswer((i) -> {
+                JacksonEvent e1 = (JacksonEvent) i.getArgument(0);
+                e1.setEventHandle(new DefaultEventHandle(acknowledgementSet1));
+                return null;
+            }).when(acknowledgementSet1).add(any(JacksonEvent.class));
+        } catch (Exception e){}
+
+        eventBuilder = mock(EventBuilder.class);
+        when(eventBuilder.withEventMetadata(any(EventMetadata.class))).thenReturn(eventBuilder);
+        when(eventBuilder.withData(any(Map.class))).thenReturn(eventBuilder);
+        when(eventBuilder.build()).thenReturn(JacksonEvent.fromEvent(event));
+        when(eventFactory.eventBuilder(EventBuilder.class)).thenReturn(eventBuilder);
+        Record firstRecord = recordsIn.iterator().next();
+        EventHandle firstHandle = ((Event)firstRecord.getData()).getEventHandle();
+        Record recordOut = getRecordStrategy.getRecord(firstRecord);
+        assertThat(recordOut, sameInstance(firstRecord));
+        assertTrue(getRecordStrategy.getReferencedRecords().contains(firstRecord));
+        recordOut = getRecordStrategy.getRecord(firstRecord);
+        assertThat(recordOut, not(sameInstance(firstRecord)));
+        assertFalse(handleRefCount.containsKey(firstHandle));
+        EventHandle newHandle = ((JacksonEvent)recordOut.getData()).getEventHandle();
+        assertTrue(getRecordStrategy.getReferencedRecords().contains(recordOut));
+        assertThat(newHandle, not(equalTo(null)));
+        assertFalse(handleRefCount.containsKey(newHandle));
+    }
+
+    @Test
+    void test_multiple_records_with_acknowledgements_and_multi_components() {
+        Collection<DataFlowComponent<PipelineConnector>> dataFlowComponents = new ArrayList<>();
+        when(pipelineDataFlowComponent.getComponent()).thenReturn(new PipelineConnector());
+        for (int i = 0; i < 3; i++) {
+            dataFlowComponents.add(pipelineDataFlowComponent);
+        }
+
+        final RouterCopyRecordStrategy getRecordStrategy = createObjectUnderTest(dataFlowComponents);
+        List<EventHandle> eventHandles = new ArrayList<>();
+        attachEventHandlesToRecordsIn(eventHandles);
+        try {
+            doAnswer((i) -> {
+                JacksonEvent e1 = (JacksonEvent) i.getArgument(0);
+                e1.setEventHandle(new DefaultEventHandle(acknowledgementSet1));
+                return null;
+            }).when(acknowledgementSet1).add(any(JacksonEvent.class));
+        } catch (Exception e){}
+
+        eventBuilder = mock(EventBuilder.class);
+        when(eventBuilder.withEventMetadata(any(EventMetadata.class))).thenReturn(eventBuilder);
+        when(eventBuilder.withData(any(Map.class))).thenReturn(eventBuilder);
+        when(eventBuilder.build()).thenReturn(JacksonEvent.fromEvent(event));
+        when(eventFactory.eventBuilder(EventBuilder.class)).thenReturn(eventBuilder);
+        Collection<Record> recordsOut = getRecordStrategy.getAllRecords(recordsIn);
+        assertThat(recordsOut.size(), equalTo(recordsIn.size()));
+
+        recordsOut = getRecordStrategy.getAllRecords(recordsIn);
+        assertThat(recordsOut.size(), equalTo(recordsIn.size()));
+
+        Set<Record> recordsOutSet =  recordsOut.stream().collect(Collectors.toSet());
+        recordsIn.forEach(recordIn -> assertFalse(recordsOutSet.contains(recordIn)));
+
+        recordsIn.stream().forEach((record) -> {
+            assertTrue(getRecordStrategy.getReferencedRecords().contains(record));
+            JacksonEvent e = (JacksonEvent)record.getData();
+            assertFalse(handleRefCount.containsKey(e.getEventHandle()));
+        });
+        recordsOut.stream().forEach((record) -> {
+            assertTrue(getRecordStrategy.getReferencedRecords().contains(record));
+            JacksonEvent e = (JacksonEvent)record.getData();
+            assertFalse(handleRefCount.containsKey(e.getEventHandle()));
+        });
+    }
 }
